feat : searchDetail 페이지 구현
