+++ source/nuxtend/pages/app/search/fullSearch/searchDetail.vue
@@ -1,0 +1,55 @@
+<template lang="html">
+  <div>
+    <h5>데이터 상세보기</h5>
+
+    <div class="component">
+      <h3>name tag component</h3>
+    </div>
+
+    <div class="component">
+      <h3>title component</h3>
+      <h3>view table component - 2</h3>
+    </div>
+
+    <div class="component">
+      <h3>title component</h3>
+      <h3>view table component 2 + 1</h3>
+    </div>
+
+    <div class="component">
+      <h3>title component</h3>
+      <h3>textarea component</h3>
+    </div>
+
+    <div class="component">
+      <h3>title component</h3>
+      <h3>tab menu component</h3>
+      <h3>button component</h3>
+      <h3>basic-table</h3>
+    </div>
+  </div>
+</template>
+
+<script type="text/javascript">
+export default {
+  name: "app-search-full-detail",
+  extends: {},
+  props: {},
+  data() {
+    return {};
+  },
+  computed: {},
+  components: {},
+  watch: {},
+  methods: {},
+  created() {}
+};
+</script>
+
+<style lang="scss">
+.component {
+  padding: 20px;
+  border: 1px solid lightgrey;
+  margin: 10px;
+}
+</style>
